Auto-generate execute methods for generated command enums
Not too long ago we added a macro to auto generate the command enums and
the required trait implementations from a concise declarative
representation. This change extends this mechanism to the execute method
implementation that some of those enums provide.
When a tuple is specified as the "destination", e.g., here:
  > Enum! {ConfigCommand, [
  >   Get => ("get", config_get),
  >   Set => ("set", config_set)
  > ]}

the second component of this tuple will be interpreted as the function
to invoke when this variant used in the newly generated execute method.