Add a script to determine the nitrocli binary size at a git revision
We have been loosely tracking the resulting size of the stripped release
binary as that is arguably the most relevant metric to optimize for. To
have a better idea of the influence of various changes and their effect
on the binary size, this change adds a script that automates the process
of gathering this metric. E.g.,
  $ var/binary-size.py HEAD~3 HEAD --unit kib
  > 994
  > 970