Check PWS slot status before accessing it
The Nitrokey devices to not check whether a PWS slot is programmed
before accessing it (upstream issues [0] [1]).  Until this is fixed in
the firmware, we have to manually check the slot status in nitrocli pws
get.  This could have been done in libnitrokey or the nitrokey crate,
yet this would lead to unnecessary commands if we check multiple fields
of a slot at the same time.

[0] https://github.com/Nitrokey/nitrokey-pro-firmware/issues/56
[1] https://github.com/Nitrokey/nitrokey-storage-firmware/issues/81