Respect ctx.std{out,err} for version output
Due to a bug in argparse [0], custom stdout and stderr settings are
ignored when using argparse::Print, as we currently do for the --version
option.  This patch adds a workaround for this problem:  Instead of
using argparse::Print, we use argparse::StoreTrue for the --version
option.

The argument parsing will fail as the command is missing, but the
version variable will still be set to true if the version option was
set.  So we ignore the parsing result and discard the argparse output if
the version variable is set.

[0] https://github.com/tailhook/rust-argparse/pull/50