Properly show the default format used in otp set subcommand
The otp set subcommand allows for three different formats in which the
user may pass in the secret, with the default being hexadecimal. By
convention we convey the default being used in the help text to the
respective command, but that default was missing here.
To that end, this change makes sure to include the default format being
used in corresponding help text.