Add note about interaction with GnuPG to README
Applications accessing the Nitrokey device through libnitrokey
apparently lock the device. This lock may not be released in time,
causing GnuPG operations performed shortly afterwards to fail (or, the
other way around, when accessing the GPG smart card through GnuPG and
then using nitrocli, the latter program may fail the interaction).
Unfortunately there is nothing we can do directly about this problem on
the nitrocli side of things, as the problem seemingly needs to be fixed
in the firmware. Hence, with this change we add a note about this
problem including a reference to the upstream issue to the README.