Update links to Debian & Ubuntu packages
This change updates the links to the Debian & Ubuntu packages in the
README. The previously used links were release-specific, which makes
them prone to becoming outdated over time.
The new Debian link now is to "stable", which is updated automatically
as time goes on and new releases are made. For Ubuntu, there is no such
meta release and so we link the search results instead, which fulfill a
similar purpose.