Use strings instead of Command variants
In one of the next patches, we will add fields to some Command variants
to be able to use them with structopt.  Then we will no longer be able
to instantiate them directly, so we replace these instances for the
transition.

This patch also removes the cmd_help! macro that is no longer needed.