Add all_str function to enums generated by Enum!
To make it easier to list all possible values for a command-line option
mapped to an enum, we add the all_str function to the Enum! macro that
returns an array of the string representations of all variants.

We also use this new function to simplify the generation of the error
message in the FromStr implementation in Enum!.