Use symbolic default values where possible
With the switch to using structopt for argument parsing some of the
default values have become mere strings, whereas one of our earlier
objectives was to have symbolic values where that is possible. As
structopt works with symbolic values equally well, this change makes it
use those.
As a bonus, also list the possible formats for the --format option to
the otp set subcommand.