Implement list command
This patch implements the list command that lists all attached Nitrokey
devices.  Currently the Nitrokey Storage does not report its serial
number during HID enumeration, see [0].  So if we detect a Nitrokey
Storage device, we connect to it and use the get_serial_number function
to query its serial number.  This can be disabled using the --no-connect
option.

Note that even the get_serial_number function reports a wrong serial
number for the Nitrokey Storage, see [1].

[0] https://github.com/Nitrokey/nitrokey-storage-firmware/issues/88
[1] https://github.com/Nitrokey/nitrokey-storage-firmware/issues/76