Use config to parse environment variables for Config
This patch uses the config crate to parse the environment.  A variable
NITROCLI_KEY can be used to overwrite the configuration for *key*.  This
has the side effect that the NITROCLI_NO_CACHE variable is evaluated as
a boolean variable (instead of only checking whether it is set).  We
also accept two new variables, NITROCLI_MODEL and NITROCLI_VERBOSITY.