Use standard configuration file locations
This patch uses the directories crate to query the appropriate path for
the configuration files.  For Linux, paths according to the XDG Base
Directory Specification are used.

Note that directories does not yet support the XDG_CONFIG_DIRS variable
for system-wide configuration files.  Therefore we only use a user
configuration file.