Add --serial-number option
This patch adds the --serial-number option that allows the user to
filter the attached Nitrokey devices by serial number.  As the Nitrokey
Storage does not include its serial number in the USB device descriptor
and as we don’t want to connect to it just to query the serial number,
this option only works for Nitrokey Storage devices.