Add config test case and example config file
This patch adds a simple configuration file that demonstrates the syntax
and contains some documentation.  We suggest to ship this file together
with nitrocli and to install it e.g., in the /usr/share/doc/nitrocli
directory.  This patch also adds a simple test case that makes sure that
the example file is parsed correctly.