Convert Cargo.lock to new lock file format
Since version 1.38.0 of Rust, Cargo has built in support for a new lock
file format that aims at reducing merge conflicts of unrelated changes.
Given that we require an even more recent version of Rust, it seems safe
to switch over to this new format (which by now is the default for new
projects anyway).
This patch changes the format of Cargo.lock accordingly.