Update nitrokey dependency to 0.7.1
This change updates the nitrokey dependency we consume to 0.7.1. With
this version the crate's Error type got reworked slightly, allowing it
to implement Sync and Send. It furthermore no longer duplicates
information available through the source() method in its Display
implementation. Both these changes will enable us to switch over to
using the anyhow crate for error handling in nitrocli in the future.