Use dedicated types for output
Previously, we just printed nitrocli’s output to the stdout stream using
format macros.  To make it possible to support multiple output formats,
this patch introduces the output module with dedicated types for the
data that is printed to stdout:
- Value:  A single value, e. g. an OTP code or the configuration data.
- Table:  A list of objects, rendered as a table, e. g. the list of OTP
  slots.

With this patch, we also drop the --quiet option for the pws get
command.  It was intended to make it easier to parse the output of the
command.  But as we will add better output formats for parsing, we no
longer need this option.

TODO: update changelog