Fail if multiple matching devices are attached
Previously, we just applied our filter (if any) to all attached Nitrokey
devices and selected the first match when connection to a Nitrokey
device.  This may lead to unexpected behavior if multiple devices are
attached.  This patch changes the find_device function to return an
error if multiple matching devices are found.