Always set model to Storage in with_storage_device
When refactoring the connect function to use the configuration struct
instead of the model and serial number options, the with_storage_device
function was broken:  If the user set the --model option, the function
still checked that it is set to Storage.  But if the user did not set
this option, the function did not make sure to only connect to a Storage
device.

This patch fixes this bug by setting the model field of the
configuration struct to Some(DeviceModel::Storage) if it is None.