Use device model instead of device in panic
with_storage_device panics if the connect method does not return a
Storage device although we set the model filter to storage.  Previously,
we used the Debug representation of the wrong device wrapper as part of
the panic message, for example:
    Pro(Pro { manager: Some(Manager { marker: PhantomData }) })
Besides the name of the model, this does not contain any useful
information.  Therefore we replace the Debug representation with the
model to make the error message more concise.