Refactor find_device into find_devices
This patch refactors the find_device function that finds a Nitrokey
device that matches the given filter or returns an error if there are
no or multiple matching Nitrokey devices into the find_devices function
that returns a vector with all matching Nitrokey devices.  The check for
exactly one matching device is now performed in the connect function.