Print SD card usage in Storage status
The Storage device keeps track of the areas of the SD card that have
been accessed during this power cycle.  This data can be accessed using
the NK_get_SD_usage_data function that returns a range of the SD card
that has not been accessed.  This data can be used as a guide line when
creating new hidden volumes.

This patch adds the SD card usage data to the output of the status
command for Nitrokey Storage devices.