Subdivide environment variable section in manual
This change reorders and subdivides the Environment section we have in
the manual. The first subsection in it is about variables pertaining the
program configuration and the second one about those influencing
password & PIN entry. Having these dedicated subsections will
subsequently allow us to reference them in follow up changes. The
reordering is meant to reflect the more general applicability that
configuration variables have.