Fix number of password safe slots mentioned in the man page
The man page mentions a password safe slot count of 20 for both the
Nitrokey Pro and the Nitrokey Storage devices. This count is incorrect,
as both devices can store only 16 entries each, as is evident from the
corresponding technical details for each device. With this change we
correct said number.