Update man page after change to base32 default OTP format
When we switched the default OTP format from hexadecimal to base32 we
missed that the manual was providing examples of the otp set command
that were implicitly making use of the default format -- leading to a
mismatch between what the example is meant to do and what it actually
does.
This change fixes this oversight in the manual by adjusting the examples
accordingly.