Set nitrocli as the default binary to run
For a while now the crate was comprised of two binaries: nitrocli itself
as well as the shell-complete support utility program. When using cargo
run on such a crate, it will seize to work and ask for advice on which
binary to run.
From a user's perspective the answer is obvious: by default it should
always be nitrocli. With this change we convey this information to Cargo
as well.