Add support for testing on Librem Key
This change adds support for running our test suite on Librem Key
devices as well. So far, nitrokey-test, the crate that we use for
dispatching tests to the individual devices, only supported Nitrokey Pro
and Storage devices. With version 0.5 it is capable of testing Librem
Key devices as well. To that end, this change updates the version of
nitrokey-test that we consume to 0.5.